requirements.txt
spacy
pandas
psycopg2-binary
